img_vectorize and path_viewer
